Removed content from the documentation that is now part of the web chat documentation site.
