--- src/common/types/errors.ts
@@ -1,95 +1,0 @@
-/**
- * (C) Copyright IBM Corp. 2020.
- *
- * Licensed under the MIT License (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * https://opensource.org/licenses/MIT
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- *
- */
-
-/**
- * All the types to manage error handling from the service desk are here. We currently support errors when connecting
- * and errors that can happen that cause the session to disconnect.
- */
-
-/**
- * The possible events that may have some form of error status.
- */
-enum ErrorType {
-  /**
-   * This error is meant to be displayed while the user is attempting to connect to a service desk and before an
-   * agent has joined. If this error is generated by the service desk, it is expected that the service desk will
-   * treat the chat as having ended (or never started).
-   */
-  CONNECTING = 1,
-
-  /**
-   * This is used to indicate the state of errors that can happen any time during a chat where the service desk
-   * implementation has lost a connection to the back-end. If this error occurs while the user is waiting for an
-   * agent to join, it will be treated as a {@link CONNECTING} error instead.
-   */
-  DISCONNECTED = 2,
-}
-
-/**
- * The type for the information passed to {@link ServiceDeskCallback#setErrorStatus}. It is a discriminating union
- * where the {@link #type} property is the discriminating value that determines which child interface is to be used.
- */
-type ServiceDeskErrorInfo = ConnectingErrorInfo | DisconnectedErrorInfo;
-
-/**
- * This is the parent interface for the information passed to {@link ServiceDeskCallback#setErrorStatus}. It is used
- * as a discriminating union where the {@link #type} property is the discriminating value that determines which
- * child interface is to be used.
- */
-interface BaseErrorInfo {
-  /**
-   * An optional value that will be logged to the console as an error.
-   */
-  logInfo?: unknown;
-}
-
-/**
- * This error is meant to be displayed while the user is attempting to connect to a service desk and before an
- * agent has joined. If this error is generated by the service desk, it is expected that the service desk will
- * treat the chat as having ended (or never started).
- */
-interface ConnectingErrorInfo extends BaseErrorInfo {
-  /**
-   * The discriminating value for this type.
-   */
-  type: ErrorType.CONNECTING;
-
-  /**
-   * An optional message that is displayed to the user in the bot view. If this value is not provided, a default
-   * message will be shown instead.
-   *
-   * @since Web chat 6.7.0. This value will be ignored if used with earlier versions of web chat.
-   */
-  messageToUser?: string;
-}
-
-/**
- * This is used to indicate the state of errors that can happen any time during a chat where the service desk
- * implementation has lost a connection to the back-end. If this error occurs while the user is waiting for an
- * agent to join, it will be treated as a {@link CONNECTING} error instead.
- */
-interface DisconnectedErrorInfo extends BaseErrorInfo {
-  /**
-   * The discriminating value for this type.
-   */
-  type: ErrorType.DISCONNECTED;
-
-  /**
-   * Indicates if the service desk has become disconnected. A value of true can be passed that will indicate that a
-   * previous disconnection is over and the service desk is now connected again.
-   */
-  isDisconnected: boolean;
-}
-
-export { ServiceDeskErrorInfo, ErrorType };
