+++ src/example/webChat/client/src/exampleServiceDesk.ts
@@ -12,16 +12,16 @@
  *
  */
 
-import { ErrorType } from 'common/types/errors';
-import { MessageRequest, MessageResponse } from 'common/types/message';
-import { User } from 'common/types/profiles';
-import {
+import { ErrorType , MessageRequest, MessageResponse , User ,
   ServiceDesk,
   ServiceDeskFactoryParameters,
   ServiceDeskStateFromWAC,
   StartChatOptions,
-} from 'common/types/serviceDesk';
-import { AgentProfile, ServiceDeskCallback } from 'common/types/serviceDeskCallback';
+, AgentProfile, ServiceDeskCallback } from 'common/types/serviceDeskTypes';
+
+
+
+
 import { stringToMessageResponseFormat } from 'common/utils';
 
 /**
