+++ src/oracle/webChat/client/src/buildEntry.ts
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  *
  */
-import { ServiceDesk, ServiceDeskFactoryParameters } from 'common/types/serviceDesk';
+import { ServiceDesk, ServiceDeskFactoryParameters } from 'common/types/serviceDeskTypes';
 
 import { OracleB2CServiceDesk } from './oracleB2CServiceDesk';
